Created Router, Signup & Login Process were improved
